+++ src/terraform_manager/terraform/versions.py
@@ -53,6 +53,7 @@ def patch_versions(
     workspaces: List[Workspace],
     *,
     new_version: str,
+    no_tls: bool = False,
     write_output: bool = False
 ) -> bool:
     """
@@ -63,6 +64,7 @@ def patch_versions(
     :param organization: The organization containing the workspaces to patch.
     :param workspaces: The workspaces to patch.
     :param new_version: The new Terraform version to assign to the workspaces.
+    :param no_tls: Whether to use SSL/TLS encryption when communicating with the Terraform API.
     :param write_output: Whether to print a tabulated result of the patch operations to STDOUT.
     :return: Whether all patch operations were successful. If even a single one failed, returns
              False.
@@ -92,6 +94,7 @@ def patch_versions(
         json=json,
         on_success=on_success,
         on_failure=on_failure,
+        no_tls=no_tls,
         write_output=write_output
     )